requirements.txt
bcrypt==4.0.1
blinker==1.6.3
boto3==1.28.83
botocore==1.31.83
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
config==0.5.1
decorator==4.4.2
env==0.1.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
moviepy==1.0.3
numpy==1.26.1
Pillow==10.0.1
proglog==0.1.10
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.9
pyasn1==0.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.22
tinytag==1.10.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
